Rename db folder to clarify internal types
