+++ src/db/queries/ReimbursementItemBoxQuery.ts
@@ -1,7 +1,9 @@
-import {ItemBoxId} from "../../types/db/public/ItemBox";
-import {ReimbursementId} from "../../types/db/public/Reimbursement";
-import ReimbursementItemBox,
-{ReimbursementItemBoxInitializer, ReimbursementItemBoxMutator} from "../../types/db/public/ReimbursementItemBox";
+import {ItemBoxId} from "../../types/db_internal/public/ItemBox";
+import {ReimbursementId} from "../../types/db_internal/public/Reimbursement";
+import ReimbursementItemBox, {
+	ReimbursementItemBoxInitializer,
+	ReimbursementItemBoxMutator
+} from "../../types/db_internal/public/ReimbursementItemBox";
 import {CompositeCrudQueryable} from "../CompositeCrudQueryable";
 import * as DB from "../../db/DB";
