+++ src/db/queries/TransactionQuery.ts
@@ -1,6 +1,6 @@
 
 import Transaction,
-{TransactionId, TransactionInitializer, TransactionMutator} from "../../types/db/public/Transaction";
+{TransactionId, TransactionInitializer, TransactionMutator} from "../../types/db_internal/public/Transaction";
 import {SimpleCrudQueryable} from "../SimpleCrudQueryable";
 
 const tableName = "transaction";
