+++ src/types/db_internal/public/ItemIndividual.ts
