+++ src/types/db_internal/public/TransactionItem.ts
