+++ src/types/db_internal/public/ValidCategory.ts
